Add method to set context in PowertoolsConfigurations.cs. Feature in Metrics. New PowertoolsEnvironment
