+++ libraries/src/AWS.Lambda.Powertools.Common/Core/PowertoolsConfigurations.cs
@@ -184,4 +184,10 @@ public class PowertoolsConfigurations : IPowertoolsConfigurations
     /// <value><c>true</c> if [tracing is disabled]; otherwise, <c>false</c>.</value>
     public bool TracingDisabled =>
         GetEnvironmentVariableOrDefault(Constants.TracingDisabledEnv, false);
+
+    /// <inheritdoc />
+    public void SetExecutionEnvironment<T>(T type)
+    {
+        _systemWrapper.SetExecutionEnvironment(type);
+    }
 }
