+++ libraries/tests/AWS.Lambda.Powertools.Common.Tests/Core/PowertoolsEnvironmentTest.cs
@@ -36,6 +36,9 @@ public class PowertoolsEnvironmentTest
     }
 }
 
+/// <summary>
+/// Fake Environment for testing
+/// </summary>
 class MockEnvironment : IPowertoolsEnvironment
 {
     private readonly Dictionary<string, string> _mockEnvironment = new();
